requirements.txt
six
lxml
requests
github3.py==0.9.3
passlib
pystache
